Move doughnut tests into integration test dir
